requirements.txt
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.10.4
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.8.0
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
openai==1.2.2
pydantic==2.4.2
pydantic_core==2.10.1
python-telegram-bot==20.6
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
